# [c++] Optimize Varint Codeing for c++