optimize Varint Codeing by BMI2 instruction. In our funciton level benchmarking, we get about 2.7x  maximum performance
    boost.
2022-10-12T09:50:01+08:00
Running ./zz_bench
Run on (128 X 3300 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x64)
  L1 Instruction 32 KiB (x64)
  L2 Unified 1280 KiB (x64)
  L3 Unified 49152 KiB (x2)
Load Average: 0.20, 0.50, 0.47
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
BM_Code_Varint32_1             3.13 ns         3.13 ns    223111504
BM_Code_BMI2_Varint32_1        3.17 ns         3.17 ns    220073339
BM_Code_Varint32_2             3.98 ns         3.98 ns    176472207
BM_Code_BMI2_Varint32_2        3.58 ns         3.58 ns    196546601
BM_Code_Varint32_3             5.86 ns         5.86 ns    119058237
BM_Code_BMI2_Varint32_3        4.42 ns         4.41 ns    158287666
BM_Code_Varint32_4             8.44 ns         8.43 ns     82981161
BM_Code_BMI2_Varint32_4        8.78 ns         8.77 ns     79598740
BM_Code_Varint32_5             10.7 ns         10.7 ns     65171356
BM_Code_BMI2_Varint32_5        7.73 ns         7.73 ns     90640481
BM_Code_Varint32_6             10.7 ns         10.7 ns     65157143
BM_Code_BMI2_Varint32_6        7.72 ns         7.71 ns     90778985
BM_Code_Varint32_7             10.7 ns         10.7 ns     65259369
BM_Code_BMI2_Varint32_7        7.73 ns         7.73 ns     90555467
BM_Code_Varint32_8             10.7 ns         10.7 ns     65215212
BM_Code_BMI2_Varint32_8        7.72 ns         7.72 ns     90833824
BM_Code_Varint32_9             10.7 ns         10.7 ns     65180833
BM_Code_BMI2_Varint32_9        7.73 ns         7.73 ns     90581863
BM_Code_Varint32_10            10.7 ns         10.7 ns     65177312
BM_Code_BMI2_Varint32_10       7.73 ns         7.72 ns     90671029
BM_Code_Varint64_1             3.20 ns         3.20 ns    218468799
BM_Code_BMI2_Varint64_1        3.18 ns         3.18 ns    220401104
BM_Code_Varint64_2             3.95 ns         3.95 ns    177792984
BM_Code_BMI2_Varint64_2        3.69 ns         3.69 ns    189878208
BM_Code_Varint64_3             6.04 ns         6.04 ns    114832150
BM_Code_BMI2_Varint64_3        4.26 ns         4.26 ns    164892413
BM_Code_Varint64_4             8.43 ns         8.43 ns     83087712
BM_Code_BMI2_Varint64_4        8.71 ns         8.71 ns     80103738
BM_Code_Varint64_5             10.7 ns         10.7 ns     65192875
BM_Code_BMI2_Varint64_5        8.68 ns         8.67 ns     80551592
BM_Code_Varint64_6             13.2 ns         13.2 ns     53063208
BM_Code_BMI2_Varint64_6        8.69 ns         8.68 ns     80423826
BM_Code_Varint64_7             15.8 ns         15.8 ns     44324556
BM_Code_BMI2_Varint64_7        8.55 ns         8.55 ns     82274517
BM_Code_Varint64_8             18.3 ns         18.3 ns     38198496
BM_Code_BMI2_Varint64_8        8.35 ns         8.35 ns     84413733
BM_Code_Varint64_9             20.8 ns         20.8 ns     33624867
BM_Code_BMI2_Varint64_9        7.75 ns         7.75 ns     90452908
BM_Code_Varint64_10            23.3 ns         23.3 ns     30076428
BM_Code_BMI2_Varint64_10       8.33 ns         8.33 ns     83756807

in our object level test. End to End benchmarking for object that all are varint fields ,we can get about 20% performance boost