@jorgbrown 2022-10-12
Thanks for taking the time to look into this, and, like last time, special thanks for taking the time to do some benchmarks.

Alas, similar to the effort to make parsing faster, you have similar caveats here: the benchmarks can be misleading because they're not tracking random data, and PDEP/PEXT can be much slower than the equivalent integer instructions on Zen and Zen2 CPUs.

But the far bigger concern with this CL is this part of the code:

>    uint64_t* pdata = reinterpret_cast<uint64_t*>(ptr);
>    if (efftiveBitCounts <= 56) {
>      *pdata = _pdep_u64(value, mask);

Here, you're writing 8 bytes of data to the output stream, even if you need fewer.  This is a problem because unlike the input streams, there is no guaranteed buffer area on the output stream.  We haven't found it to be significantly beneficial to allocate more buffer area at the very end of the output stream, so we don't.  You have to take care not to write any more data than is actually needed; if you write more, you may be overwriting some other buffer!