@zzachimed 2022-11-19
@jorgbrown 👍
verry thanks and appreciate for your carefuly comments . sorry for latter response for current busy work.

1 for your concern about avx512 issue
we disucssed with intel senior principal architect Moga, Adrian C [[adrian.c.moga@intel.com](mailto:adrian.c.moga@intel.com)]. This is his comment about"
Intel can slow down the clock rate of every core on a chip just because one of them started using AVX-512"
"I don’t this this is true.
First, the article is old and based on a SKX 12C SKU.

Second, the Cloudflare blog writer is confused by wikichip reference, which refers to N cores using the same license.
Nothing in there says that one core doing AVX-512 brings down the other cores to base frequency if the run SSE.

Also, if I understand his experiment correctly, he is running a fully-loaded system with two deployments.
So, he’s observing a noisy neighbor effect for frequency under TDP constraints.
If some cores use more power with AVX-512, the other cores may also see a frequency degradation.
I think even that may not be true starting with SPR, as the cores have autonomous power management.
But I am not 100% certain. I will ask and let you know.

Thanks!
Adrian
"
2 for your concern about run-out memory issue "
But the far bigger concern with this CL is this part of the code:

uint64_t* pdata = reinterpret_cast<uint64_t*>(ptr);
if (efftiveBitCounts <= 56) {
*pdata = _pdep_u64(value, mask);

Here, you're writing 8 bytes of data to the output stream, 
":
current stream mechinism can additional bytes for our processing . our decode process varintparse is private function and server for up message level depend on the buffer from streaming processing. In sepecifct words the buffer is from EpsCopyOutputStream:: EnsureSpace processing(every object encode start it will call this func). Maybe you could deep dive in this process you will found some wondeful mechnism to prevent memory run out in some range .In this buffer size check and buffer swap processing ,there has some woderful mechinisim to processing some visit out of range issue --at most additional 16 bytes (kSlopBytes)for run-out processing(in EpsCopyOutputStreamclass ) . It afford additional in "char buffer_[2 * kSlopBytes] = {};" . The specific processing is comlicated and hard to description very shortly. So its better to check that ParseContext::Done code line by line.