@zzachimed 2022-11-21
@jorgbrown 👍
verry thanks and appreciate for your carefuly comments . sorry for latter response for current busy work.
1     for your concern about pext execution effiency issue in AMD platform “PDEP/PEXT can be much slower than the equivalent integer instructions on Zen and Zen2 CPUs.”:

       It looks like Zen3 will have some improvement regarding to this.


2    for your concern about run-out memory issue ":
But the far bigger concern with this CL is this part of the code:
uint64_t* pdata = reinterpret_cast<uint64_t*>(ptr);
if (efftiveBitCounts <= 56) {
*pdata = _pdep_u64(value, mask);
Here, you're writing 8 bytes of data to the output stream,
":
                current stream mechinism can additional bytes for our processing . our encode process UnsafeVarint is private function and server for up message level depend on the buffer from streaming processing. In sepecifct words the buffer is from EpsCopyOutputStream:: EnsureSpace processing(every object encode start it will call this func). Maybe you could deep dive in this process you will found some wondeful mechnism to prevent memory run out in some range .In this buffer size check and buffer swap processing ,there has some woderful mechinisim to processing some visit out of range issue --at most additional 16 bytes (kSlopBytes)for run-out processing(in EpsCopyOutputStreamclass ) . It afford additional space in "char buffer_[2 * kSlopBytes] = {};" . The specific processing is comlicated and hard to description very shortly. So its better to check that ParseContext::Done code line by line.