@jorgbrown 2023-01-07
It's true that EpsCopyOutputStream has the capability to use a separate buffer, which would allow you to use your faster BMI routines.

Unfortunately it also has this constructor:
```
// Only for array serialization. No overflow protection, end_ will be the
// pointed to the end of the array. When using this the total size is already
// known, so no need to maintain the slop region.
EpsCopyOutputStream(void* data, int size, bool deterministic)
    : end_(static_cast<uint8_t*>(data) + size),
      buffer_end_(nullptr),
      stream_(nullptr),
      is_serialization_deterministic_(deterministic) {}
```
This is one we use quite often at Google, and I believe is the one used when serializing directly into a std::string.  And since it does not maintain a slop buffer, you can't write even one byte more than you actually need to.