@jorgbrown 2023-01-10
I'd be curious how your benchmarks come out if you use this for your PDEP; it works on all CPUs:
```
inline uint64_t handPDEP_7F7F7F7F7F7F7F7F(uint64_t value) {
  // 1. Reshape value into two halves, each of which have 28 bits
#if defined(__aarch64__)
  uint64_t top28 = value & 0x00FFFFFFF0000000;
  uint64_t bot28 = value & 0x000000000FFFFFFF;
  value = (top28 << 4) + bot28;
#else
  value <<= 4;
  asm("" : "+r"(value));  // Improve clang
  value = (value & 0x0FFFFFFF00000000) | (static_cast<uint32_t>(value) >> 4);
#endif
  // 2. Reshape value into four quarters, each of which have 14 bits
  uint64_t top14s = value & 0xFFFFC000'0FFFC000;
  uint64_t bot14s = value - top14s;
  value = bot14s + (top14s << 2);
  // 3. Reshape value into eight bytes, each of which have 7 bits
  uint64_t top7s = value & 0x3F80'3F80'3F80'3F80;
  value += top7s;
  return value;
}
```